deep_learning_model/requirements.txt
Sphinx
coverage
#awscli
flake8
python-dotenv>=0.5.1
